python-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
Pillow==11.0.0
opencv-python-headless==4.10.0.84
numpy>=1.26.0
python-docx==1.1.2
reportlab==4.2.5
ebooklib==0.18.1
ffmpeg-python==0.2.0
python-multipart==0.0.12
